dataset generator/requirements.txt
openai
langchain
streamlit==1.28.2
watchdog
requests
pandas
python-dotenv
tiktoken
docx2txt
pypdf
PyPDF2
aiohttp
PyYAML
beautifulsoup4
openpyxl
unstructured
networkx
llama-index
langchain-community